[dv,top] Share common code in pwrmgr_*sleep_all_wake_ups
This facilitates the creation of a random version of these tests,
and removes redundant code.
